pyproject.toml
[project]
name = "openai-moss-agents"
version = "0.1.0-dev"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "thirdgerb", email = "thirdgerb@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "ghostos-moss>=0.3.7",
    "openai-agents>=0.0.3",
]

[project.scripts]
openai-moss-agents = "openai_moss_agents:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
console = [
    "rich>=13.9.4",
]

